bug de modulo ruta reparado

+++ SistemaEncomiendas/aplicaciones/moduloEntRec/views.py
@@ -5,13 +5,13 @@ from .models import Ruta
 from .forms import RutaForm
 from ..moduloUsuarios.models import Repartidor
 from django.urls import reverse_lazy
-from django.views.generic import CreateView,ListView,UpdateView,DeleteView
+from django.views.generic import CreateView,ListView,UpdateView,DeleteView,DetailView
 # Create your views here.
 
 class ListarRuta(ListView):
     model = Ruta
     template_name = 'moduloEntRec/listarRutas.html'
-    context_object_name = 'ruta'
+    context_object_name = 'rutas'
 class CrearRuta(CreateView):
     model = Ruta
     form_class = RutaForm
@@ -20,7 +20,7 @@ class CrearRuta(CreateView):
 
     def get_form(self, *args, **kwargs):
         form = super().get_form(*args, **kwargs)
-        form.fields['repartidor'].queryset = Repartidor.objects.all() 
+        form.fields['id_repartidor'].queryset = Repartidor.objects.all() 
         return form
     
 class EditarRuta(UpdateView):
@@ -28,22 +28,33 @@ class EditarRuta(UpdateView):
     form_class = RutaForm
     template_name = 'moduloEntRec/editarRuta.html'
     success_url = reverse_lazy('moduloEntRec:ListarRutas')
+    
+    def get_form(self, *args, **kwargs):
+        form = super().get_form(*args, **kwargs)
+        form.fields['id_repartidor'].queryset = Repartidor.objects.all() 
+        return form
 
+class VerRuta(DetailView):
+    model=Ruta
+    template_name='moduloEntRec/verRuta.html'
+    context_object_name='rutas'
+    
 
-class EliminarRuta(DeleteView):
-    model = Ruta
-    template_name = "moduloEntRec/listarRuta.html"
-    success_url = reverse_lazy('moduloEntRec:ListarRutas')
+def EliminarRuta(request, pk):
+    ruta = get_object_or_404(Ruta, pk=pk)
 
-    def delete(self, request, *args, **kwargs):
-        ruta = self.get_object()
+    if request.method == "POST":
         try:
             ruta.delete()
+            # Agregar mensaje de éxito
             messages.success(request, 'Ruta eliminada correctamente!')
         except Exception as e:
+            # Agregar mensaje de error si ocurre una excepción
             messages.error(request, f'Error al eliminar la ruta: {str(e)}')
 
         if request.headers.get('x-requested-with') == 'XMLHttpRequest':
             return JsonResponse({'message': 'Ruta eliminada correctamente!'})
-        
-        return super().delete(request, *args, **kwargs)
+        return redirect("moduloEntRec:ListarRutas")
+
+    return render(request, "moduloEntRec/eliminarRuta.html", {"ruta": ruta})
+
